refactor(insights): extract donor window-activity signal (NPPD-1696)
